remove infra folder
As much as possible, we're going to aim for no infra in this project.

In the very short term, we're going to rely on the CI nodes from the
daml repo, but moving forward we'll try to switch to CircleCI as soon as
we're on sbt.

CHANGELOG_BEGIN
CHANGELOG_END