clean up YML files
All we really want to do for now is keep building the JSON API to
double-check we're not removing more than needed.

Ability to release will be reintroduced later on when we switch to SBT.

We also don't need all of the other jobs from the daml repo that are not
involved in building PRs (weekly, daily, etc.).